Getting closer to final touch to Planetary Interaction monitor.
Former-commit-id: 73cb778b652572d4518c2dc3c5443b960a69c3a1

+++ EVEMon.Common/Serialization/API/SerializablePlanetaryPin.cs
@@ -9,13 +9,13 @@ namespace EVEMon.Common.Serialization.API
         public long PinID { get; set; }
 
         [XmlAttribute("typeID")]
-        public long TypeID { get; set; }
+        public int TypeID { get; set; }
         
         [XmlAttribute("typeName")]
         public string TypeName { get; set; }
 
         [XmlAttribute("schematicID")]
-        public long SchematicID { get; set; }
+        public int SchematicID { get; set; }
 
         [XmlAttribute("cycleTime")]
         public short CycleTime { get; set; }
@@ -24,7 +24,7 @@ namespace EVEMon.Common.Serialization.API
         public int QuantityPerCycle { get; set; }
 
         [XmlAttribute("contentTypeID")]
-        public long ContentTypeID { get; set; }
+        public int ContentTypeID { get; set; }
 
         [XmlAttribute("contentTypeName")]
         public string ContentTypeName { get; set; }
